Possibly more secure means of getting external IP
I was a little concerned returning CURL output to the command line, in the event the host in the script could somehow become compromised, partially because I am considering buffering the output of this script into a JSON string that I can push to my own API and the parse on a webpage using JavaScript. I swapped it out to us a dig command that seems to work well enough and feels more secure to me. Just thought I'd share, no pressure to merge it in. Thanks for the great script!